Added frontend

+++ backend/controlers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -20,6 +21,7 @@ const setGoals = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
+        user: req.user.id,
         text: req.body.text
     })
     res.status(200).json(goal)
@@ -36,6 +38,18 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
 
     res.status(200).json(updatedGoal)
@@ -50,6 +64,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new error('User not found')
+    }
+
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     await Goal.findByIdAndDelete(req.params.id)
